Change to use the functions from  model_wrappers.py
In this version, the config file is used to select which model will run. In production
where everything runs on a scheduler, both models will run.

This will require some improvements to the data ingest, whereby we check for the
presence of new data instead of ingesting all of the pre-processing data every
time the model is run.